Añadido el login y bootstrap a la aplicación

+++ app/Http/Controllers/EmpleadoController.php
@@ -48,7 +48,9 @@ class EmpleadoController extends Controller
 
         Empleado::insert($datosEmpleado);
 
-        return response()->json($datosEmpleado);
+        //return response()->json($datosEmpleado);
+
+        return redirect('empleado')->with('mensaje','Empleado insertado con éxito.');
     }
 
     /**
@@ -107,8 +109,12 @@ class EmpleadoController extends Controller
      */
     public function destroy($id)
     {
-        Empleado::destroy($id);
-
-        return redirect('empleado');
+        $empleado = Empleado::findOrFail($id);
+        if(Storage::delete('public/'.$empleado->Foto))
+        {
+            Empleado::destroy($id);
+        }        
+        
+        return redirect('empleado')->with('mensaje','Empleado borrado con éxito.');
     }
 }
